+++ upcloud/load_balancer.go
@@ -18,6 +18,8 @@ type (
 	LoadBalancerCertificateBundleType             string
 	LoadBalancerProxyProtocolVersion              string
 	LoadBalancerHealthCheckType                   string
+	LoadBalancerNetworkType                       string
+	LoadBalancerAddressFamily                     string
 )
 
 const (
@@ -101,6 +103,12 @@ const (
 
 	LoadBalancerHealthCheckTypeTCP  LoadBalancerHealthCheckType = "tcp"
 	LoadBalancerHealthCheckTypeHTTP LoadBalancerHealthCheckType = "http"
+
+	LoadBalancerNetworkTypePublic  LoadBalancerNetworkType   = "public"
+	LoadBalancerNetworkTypePrivate LoadBalancerNetworkType   = "private"
+	LoadBalancerAddressFamilyIPv4  LoadBalancerAddressFamily = "IPv4"
+	// IPv6 addresses are not supported yet
+	// LoadBalancerAddressFamilyIPv6  LoadBalancerAddressFamily = "IPv6"
 )
 
 // LoadBalancerPlan represents load balancer plan details
@@ -115,6 +123,7 @@ type LoadBalancerFrontend struct {
 	Name           string                          `json:"name,omitempty"`
 	Mode           LoadBalancerMode                `json:"mode,omitempty"`
 	Port           int                             `json:"port,omitempty"`
+	Networks       []LoadBalancerFrontendNetwork   `json:"networks,omitempty"`
 	DefaultBackend string                          `json:"default_backend,omitempty"`
 	Rules          []LoadBalancerFrontendRule      `json:"rules,omitempty"`
 	TLSConfigs     []LoadBalancerFrontendTLSConfig `json:"tls_configs,omitempty"`
@@ -203,13 +212,15 @@ type LoadBalancer struct {
 	Name             string                       `json:"name,omitempty"`
 	Zone             string                       `json:"zone,omitempty"`
 	Plan             string                       `json:"plan,omitempty"`
-	NetworkUUID      string                       `json:"network_uuid,omitempty"`
-	DNSName          string                       `json:"dns_name,omitempty"`
+	NetworkUUID      string                       `json:"network_uuid,omitempty"` // deprecated
+	Networks         []LoadBalancerNetwork        `json:"networks,omitempty"`
+	DNSName          string                       `json:"dns_name,omitempty"` // deprecated
 	ConfiguredStatus LoadBalancerConfiguredStatus `json:"configured_status,omitempty"`
 	OperationalState LoadBalancerOperationalState `json:"operational_state,omitempty"`
 	Frontends        []LoadBalancerFrontend       `json:"frontends,omitempty"`
 	Backends         []LoadBalancerBackend        `json:"backends,omitempty"`
 	Resolvers        []LoadBalancerResolver       `json:"resolvers,omitempty"`
+	Nodes            []LoadBalancerNode           `json:"nodes,omitempty"`
 	CreatedAt        time.Time                    `json:"created_at,omitempty"`
 	UpdatedAt        time.Time                    `json:"updated_at,omitempty"`
 }
@@ -324,4 +335,32 @@ type LoadBalancerCertificateBundle struct {
 
 	Type             LoadBalancerCertificateBundleType             `json:"type,omitempty"`
 	OperationalState LoadBalancerCertificateBundleOperationalState `json:"operational_state,omitempty"`
+}
+
+// LoadBalancerNetwork represents network attached to loadbalancer
+type LoadBalancerNetwork struct {
+	UUID        string                    `json:"uuid,omitempty"`
+	Name        string                    `json:"name,omitempty"`
+	Type        LoadBalancerNetworkType   `json:"type,omitempty"`
+	Family      LoadBalancerAddressFamily `json:"family,omitempty"`
+	IPAddresses []LoadBalancerIPAddress   `json:"ip_addresses,omitempty"`
+	DNSName     string                    `json:"dns_name,omitempty"`
+	CreatedAt   time.Time                 `json:"created_at,omitempty"`
+	UpdatedAt   time.Time                 `json:"updated_at,omitempty"`
+}
+
+// LoadBalancerIPAddress represents IP address inside loadbalancer service
+type LoadBalancerIPAddress struct {
+	Address string `json:"address,omitempty"`
+	Listen  bool   `json:"listen"`
+}
+
+// LoadBalancerNode represents loadbalancer node
+type LoadBalancerNode struct {
+	Networks []LoadBalancerNetwork `json:"networks,omitempty"`
+}
+
+// LoadBalancerNetwork represents network attached to loadbalancer
+type LoadBalancerFrontendNetwork struct {
+	Name string `json:"name,omitempty"`
 }